// src/app/mill_management_system/model/arrival/mill-deposit-details.model.ts
export class MillDepositDetails {

    serialNo: number | undefined;
    arrivalDate: any | undefined;
    noOfBags: number = 0;
    quantity: number = 0;
    transporterName: string | undefined;
    vehicleNo: string | undefined;
    driverName: string | undefined;
    driverMobNo: string | undefined;
    deliveryChallanNo: string | undefined;
}